PR review changes
Changes as per feedback given on github
